+++ RELEASE-NOTES.md
@@ -30,6 +30,8 @@ Picocli follows [semantic versioning](http://semver.org/).
 - [#697] Option sort in usage help should ignore option name prefix; long options without short name should be inserted alphabetically, instead of always appear at the top.
 - [#695] Fix runtime warnings about illegal reflective access to field `java.io.FilterOutputStream.out`. Thanks to [gitfineon](https://github.com/gitfineon) for reporting this issue.
 - [#698] Reduce `reflect-config.json` used by GraalVM native-image generation
+- [#500] Add a generic and extensible picocli annotation processor
+- [#699] Add annotation processor that generates `reflect-config.json` during build
 
 ## <a name="4.0.0-beta-1-deprecated"></a> Deprecations